mylib updated with m_swap.c + normed code

+++ srcs/memory/m_dup.c
@@ -6,7 +6,7 @@
 /*   By: anradix <anradix@student.42.fr>            +#+  +:+       +#+        */
 /*                                                +#+#+#+#+#+   +#+           */
 /*   Created: 2020/07/10 00:28:29 by anradix           #+#    #+#             */
-/*   Updated: 2020/07/17 13:34:25 by anradix          ###   ########.fr       */
+/*   Updated: 2020/09/16 19:56:11 by anradix          ###   ########.fr       */
 /*                                                                            */
 /* ************************************************************************** */
 
@@ -14,14 +14,13 @@
 
 char	*m_dup(const char *s)
 {
-	size_t	len;
+	int		len;
 	char	*dup;
-	
+
 	len = s_len(s, 0, 0);
 	if (!s || !(dup = m_alloc(len, '\0')))
 		return (NULL);
 	while (len--)
 		dup[len] = s[len];
 	return (dup);
-
 }
